PlatformIcon and FleetIcon to typescript
