+++ frontend/components/icons/FleetIcon/FleetIcon.tests.tsx
