+++ frontend/components/icons/FleetIcon/index.ts
