+++ frontend/components/icons/PlatformIcon/PlatformIcon.tests.tsx
