+++ frontend/components/icons/PlatformIcon/index.ts
